feat: auto-generate ci and test workflow files
update-workflows script looks for supported distributions' (as of now,
Debiand and Ubuntu) folder at the root level of the repository and look
up any subfolders each have and in return generate a matrix of supported
disro/coenme. With this matrix it will regenerate:

- .github/workflows/ci.yml
- .github/workflows/test.yml

And if there is any changes to these file, it will commit the changes.

The update-workflows is set to run on pull requests, on main branch, and
nightly.